requirements.txt
flask
langchain
chromadb
beautifulsoup4
requests
python-dotenv
langchain-community